:recycle: rename init package
