+++ src/main/java/com/tale/controller/IndexController.java
@@ -5,7 +5,7 @@ import com.blade.mvc.annotation.*;
 import com.blade.mvc.http.Request;
 import com.blade.mvc.http.Response;
 import com.blade.mvc.http.Session;
-import com.tale.init.TaleConst;
+import com.tale.bootstrap.TaleConst;
 import com.tale.model.dto.Archive;
 import com.tale.model.dto.Types;
 import com.tale.model.entity.Contents;
