+++ src/main/java/com/tale/controller/admin/IndexController.java
@@ -15,7 +15,7 @@ import com.blade.mvc.http.Request;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.controller.BaseController;
 import com.tale.extension.Commons;
-import com.tale.init.TaleConst;
+import com.tale.bootstrap.TaleConst;
 import com.tale.model.dto.BackResponse;
 import com.tale.model.dto.LogActions;
 import com.tale.model.dto.Statistics;
