+++ src/main/java/com/tale/controller/admin/ThemeController.java
@@ -11,8 +11,8 @@ import com.blade.mvc.http.Request;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.controller.BaseController;
 import com.tale.extension.Commons;
-import com.tale.init.TaleConst;
-import com.tale.init.TaleLoader;
+import com.tale.bootstrap.TaleConst;
+import com.tale.bootstrap.TaleLoader;
 import com.tale.model.dto.LogActions;
 import com.tale.model.dto.ThemeDto;
 import com.tale.model.entity.Logs;
