+++ src/main/java/com/tale/extension/Commons.java
@@ -2,7 +2,7 @@ package com.tale.extension;
 
 import com.blade.kit.*;
 import com.tale.controller.BaseController;
-import com.tale.init.TaleConst;
+import com.tale.bootstrap.TaleConst;
 import com.tale.service.SiteService;
 import com.tale.utils.TaleUtils;
 import com.vdurmont.emoji.EmojiParser;
